IFilter and first implementation. #23
